+++ contracts/interfaces/IFurnace.sol
@@ -12,19 +12,19 @@ interface IFurnace is IComponent {
     // Initialization
     function init(
         IMain main_,
-        uint256 period_,
+        uint32 period_,
         int192 ratio_
     ) external;
 
     /// Emitted when the melting period is changed
     /// @param oldPeriod The old period
     /// @param newPeriod The new period
-    event PeriodSet(uint256 indexed oldPeriod, uint256 indexed newPeriod);
+    event PeriodSet(uint32 indexed oldPeriod, uint32 indexed newPeriod);
 
-    function period() external view returns (uint256);
+    function period() external view returns (uint32);
 
     /// @custom:governance
-    function setPeriod(uint256) external;
+    function setPeriod(uint32) external;
 
     /// Emitted when the melting ratio is changed
     /// @param oldRatio The old ratio
