+++ contracts/p0/Furnace.sol
@@ -13,29 +13,29 @@ contract FurnaceP0 is ComponentP0, IFurnace {
     using FixLib for int192;
 
     int192 public ratio; // {1} What fraction of balance to melt each period
-    uint256 public period; // {seconds} How often to melt
-    uint256 public lastPayout; // {seconds} The last time we did a payout
+    uint32 public period; // {seconds} How often to melt
+    uint32 public lastPayout; // {seconds} The last time we did a payout
     uint256 public lastPayoutBal; // {qRTok} The balance of RToken at the last payout
 
     function init(
         IMain main_,
-        uint256 period_,
+        uint32 period_,
         int192 ratio_
     ) public initializer {
         __Component_init(main_);
         period = period_;
         ratio = ratio_;
-        lastPayout = block.timestamp;
+        lastPayout = uint32(block.timestamp);
         lastPayoutBal = main_.rToken().balanceOf(address(this));
         require(period != 0, "period cannot be zero");
     }
 
     /// Performs any melting that has vested since last call.
     function melt() external {
-        if (block.timestamp < lastPayout + period) return;
+        if (uint32(block.timestamp) < uint64(lastPayout) + period) return;
 
         // # of whole periods that have passed since lastPayout
-        uint256 numPeriods = (block.timestamp - lastPayout) / period;
+        uint32 numPeriods = (uint32(block.timestamp) - lastPayout) / period;
 
         // Paying out the ratio r, N times, equals paying out the ratio (1 - (1-r)^N) 1 time.
         int192 payoutRatio = FIX_ONE.minus(FIX_ONE.minus(ratio).powu(numPeriods));
@@ -49,7 +49,7 @@ contract FurnaceP0 is ComponentP0, IFurnace {
     }
 
     /// Period setting
-    function setPeriod(uint256 period_) external onlyOwner {
+    function setPeriod(uint32 period_) external onlyOwner {
         require(period_ != 0, "period cannot be zero");
         emit PeriodSet(period, period_);
         period = period_;
